requirements.txt
argostranslate==1.8.0
Pillow==9.4.0
pyocr==0.8.3
pyscreenshot==3.0
PySimpleGUI==4.60.4
